Remove handleClick function from the Square and add it to the Board
Now the parent Board component passes props to the child Square components so that they can be displayed correctly.
When clicking on a Square, the child Square component now asks the parent Board component to update the state of the board.
When the Board’s state changes, both the Board component and every child Square re-renders automatically.
Keeping the state of all squares in the Board component will allow it to determine the winner in the future.